move map a folder up
