+++ frontend/src/scenes/notebooks/Nodes/NotebookNodeMap.tsx
@@ -6,10 +6,10 @@ import { personLogic } from 'scenes/persons/personLogic'
 import { useValues } from 'kea'
 import { LemonSkeleton } from '@posthog/lemon-ui'
 import { NotFound } from 'lib/components/NotFound'
-import { Map } from '../../../../lib/components/Map/Map'
-import { notebookNodeLogic } from '../notebookNodeLogic'
+import { Map } from '../../../lib/components/Map/Map'
+import { notebookNodeLogic } from './notebookNodeLogic'
 import { NotebookNodeProps } from 'scenes/notebooks/Notebook/utils'
-import { NotebookNodeEmptyState } from '../components/NotebookNodeEmptyState'
+import { NotebookNodeEmptyState } from './components/NotebookNodeEmptyState'
 
 const Component = ({ attributes }: NotebookNodeProps<NotebookNodeMapAttributes>): JSX.Element | null => {
     const { id } = attributes
